requirements.txt
# Core dependencies
torch>=1.9.0
open_clip_torch>=2.0.0
Pillow>=9.0.0
tqdm>=4.62.0
numpy>=1.20.0
requests>=2.25.0
transformers>=4.18.0

# Optional dependencies
matplotlib>=3.4.0  # For visualization
pandas>=1.3.0      # For data manipulation


